refactor(bot): split trackhandlers module into focused units
extracts three concerns from trackhandlers.ts into separate modules:
- trackhistorycache.ts: dead code cache infrastructure
- autoplayoutcometracking.ts: autoplay feedback recording
- trackeventhandlers.ts: player event handler orchestration

reorganizes tests per module and updates setup.ts cache clearing.
