+++ packages/bot/src/handlers/player/trackEventHandlers.ts
@@ -1,6 +1,5 @@
 import type { Track, GuildQueue } from 'discord-player'
 import { QueueRepeatMode } from 'discord-player'
-import { LRUCache } from 'lru-cache'
 import { infoLog, debugLog, errorLog } from '@lucky/shared/utils'
 import { addTrackToHistory } from '../../utils/music/duplicateDetection'
 import { replenishQueue } from '../../services/musicManagement/queueOperations'
@@ -28,100 +27,18 @@ import {
 } from '../../services/musicManagement/replenishSuppressionStore'
 import { handleQueueExhaustion } from './queueExhaustion'
 import { recordRecommendationOutcome } from '../../services/musicRecommendation/recommendationTelemetry'
-
-const MAX_GUILD_ENTRIES = 500
-const TRACK_STATE_TTL_MS = 30 * 60 * 1000
-
-// Autoplay recommendation outcome threshold: a track played past this fraction
-// is "accepted"; ended/skipped before it is "rejected" (symmetric across the
-// playerFinish + playerSkip paths). Tune via Phase C data.
-export const OUTCOME_ACCEPT_PLAY_RATIO = 0.3
-
-export const lastPlayedTracks = new LRUCache<string, Track>({
-    max: MAX_GUILD_ENTRIES,
-    ttl: TRACK_STATE_TTL_MS,
-    updateAgeOnGet: true,
-})
-
-// Keyed per TRACK (guildId + track id), not per guild: autoplay track
-// lifecycles overlap — discord-player can emit the next track's playerStart
-// before the previous track's playerFinish/playerSkip. A single per-guild
-// timestamp gets clobbered by that interleaving, making completionRatio ≈ 0
-// for the wrong track and corrupting the accept/reject classification (#1275).
-const trackStartTimes = new LRUCache<string, number>({
-    max: MAX_GUILD_ENTRIES,
-    ttl: TRACK_STATE_TTL_MS,
-    updateAgeOnGet: true,
-})
-
-const trackStartKey = (guildId: string, trackId: string): string =>
-    `${guildId}::${trackId}`
-
-const guildRecentSkipCounts = new LRUCache<string, number>({
-    max: MAX_GUILD_ENTRIES,
-    ttl: TRACK_STATE_TTL_MS,
-    updateAgeOnGet: true,
-})
-
-export function getRecentSkipCount(guildId: string): number {
-    return guildRecentSkipCounts.get(guildId) ?? 0
-}
-
-export type TrackHistoryEntry = {
-    url: string
-    title: string
-    author: string
-    thumbnail?: string
-    timestamp: number
-}
-
-export const recentlyPlayedTracks = new LRUCache<string, TrackHistoryEntry[]>({
-    max: MAX_GUILD_ENTRIES,
-    ttl: TRACK_STATE_TTL_MS,
-    updateAgeOnGet: true,
-})
-
-function getTrackRequesterId(track: Track): string | undefined {
-    const metadata = track.metadata as { requestedById?: string } | undefined
-    return track.requestedBy?.id ?? metadata?.requestedById
-}
-
-function isAutoplayTrack(track: Track, clientUserId?: string): boolean {
-    const metadata = track.metadata as { isAutoplay?: boolean } | undefined
-    return (
-        metadata?.isAutoplay === true || track.requestedBy?.id === clientUserId
-    )
-}
-
-function isRecommendationAutoplay(track: Track): boolean {
-    const metadata = track.metadata as { isAutoplay?: boolean } | undefined
-    return metadata?.isAutoplay === true
-}
-
-async function recordImplicitTrackFeedback(
-    track: Track,
-    type: 'implicit_like' | 'implicit_dislike',
-): Promise<void> {
-    const requesterId =
-        track.requestedBy?.id ??
-        (track.metadata as { requestedById?: string } | undefined)
-            ?.requestedById
-    if (!requesterId) return
-    const trackKey = normalizeTrackKey(track.title, track.author)
-    await recommendationFeedbackService.recordImplicitFeedback(
-        requesterId,
-        trackKey,
-        type,
-    )
-}
-
-function evictOldEntries(): void {
-    for (const [guildId, entries] of recentlyPlayedTracks.entries()) {
-        if (entries.length > MAX_GUILD_ENTRIES) {
-            recentlyPlayedTracks.set(guildId, entries.slice(-MAX_GUILD_ENTRIES))
-        }
-    }
-}
+import { evictOldEntries } from './trackHistoryCache'
+import {
+    getRecentSkipCount,
+    isAutoplayTrack,
+    isRecommendationAutoplay,
+    recordImplicitTrackFeedback,
+    logAutoplayOutcomeEval,
+    trackStartTimes,
+    trackStartKey,
+    guildRecentSkipCounts,
+    OUTCOME_ACCEPT_PLAY_RATIO,
+} from './autoplayOutcomeTracking'
 
 type PlayerEvents = {
     events: { on: (event: string, handler: Function) => void }
@@ -176,6 +93,11 @@ function handleAutoplayCounter(
             message: `Manual track played but autoplay is enabled - keeping autoplay counter for radio experience`,
         })
     }
+}
+
+function getTrackRequesterId(track: Track): string | undefined {
+    const metadata = track.metadata as { requestedById?: string } | undefined
+    return track.requestedBy?.id ?? metadata?.requestedById
 }
 
 async function handleQueueReplenishment(
@@ -302,48 +224,6 @@ async function scrobbleAndRecord(
     if (!trackToRecord) return
     await scrobbleCurrentTrackIfLastFm(queue, trackToRecord)
     await addTrackToHistory(trackToRecord, queue.guild.id)
-}
-
-// #1275 diagnostic: the per-event accept/reject logic is correct and
-// unit-tested (incl. the interleaving probe), yet prod records 0 rejected.
-// The existing "Track skipped" log lacks the fields to tell a code issue
-// (missing start time, skips routing through playerFinish, the real skipRatio
-// distribution) from a genuinely-rare signal (most picks are over-queued and
-// never played → 'pending'). Emit the decision inputs for every autoplay
-// terminal event, on both paths, so Loki can disambiguate H1 vs H2.
-const logAutoplayOutcomeEval = (
-    path: 'finish' | 'skip',
-    queue: GuildQueue,
-    track: Track,
-    startTime: number | undefined,
-): void => {
-    const playedRatio =
-        startTime !== undefined && track.durationMS
-            ? (Date.now() - startTime) / track.durationMS
-            : null
-    const recordedOutcome =
-        playedRatio === null
-            ? 'none(no-timing)'
-            : playedRatio < OUTCOME_ACCEPT_PLAY_RATIO
-              ? 'rejected'
-              : path === 'finish'
-                ? 'accepted'
-                : 'ambiguous(dropped)'
-    infoLog({
-        message: 'Autoplay outcome eval',
-        data: {
-            path,
-            guildId: queue.guild.id,
-            trackId: track.id,
-            hasStartTime: startTime !== undefined,
-            durationMS: track.durationMS ?? null,
-            playedRatio:
-                playedRatio === null
-                    ? null
-                    : Math.round(playedRatio * 1000) / 1000,
-            recordedOutcome,
-        },
-    })
 }
 
 const handlePlayerFinish = async (
