+++ packages/bot/tests/setup.ts
@@ -116,14 +116,17 @@ async function clearAllCaches() {
     }
 
     try {
-        const {
-            lastPlayedTracks,
-            recentlyPlayedTracks,
-            trackStartTimes,
-            guildRecentSkipCounts,
-        } = await import('../src/handlers/player/trackHandlers')
+        const { lastPlayedTracks, recentlyPlayedTracks } =
+            await import('../src/handlers/player/trackHistoryCache')
         lastPlayedTracks.clear()
         recentlyPlayedTracks.clear()
+    } catch {
+        // May not be loaded
+    }
+
+    try {
+        const { trackStartTimes, guildRecentSkipCounts } =
+            await import('../src/handlers/player/autoplayOutcomeTracking')
         trackStartTimes.clear()
         guildRecentSkipCounts.clear()
     } catch {
